feat: Update space default avatar when renaming space - MEED-3278 - Meeds-io/MIPs#83 (#3449)
This PR allows to map the space profile full name with the space#displayName in order to compare them and delete the default space avatar and generate a new one based on updated space name value.

+++ component/core/src/main/java/org/exoplatform/social/core/space/SpaceUtils.java
@@ -464,18 +464,22 @@ public class SpaceUtils {
   public static void updateDefaultSpaceAvatar(Space space) {
     IdentityManager identityManager = CommonsUtils.getService(IdentityManager.class);
     Identity spaceIdentity = identityManager.getOrCreateSpaceIdentity(space.getPrettyName());
-    FileItem spaceAvatar = identityManager.getAvatarFile(spaceIdentity);
-    FileService fileService = CommonsUtils.getService(FileService.class);
-    if (spaceAvatar != null && spaceAvatar.getFileInfo().getId() != null
-        && EntityConverterUtils.DEFAULT_AVATAR.equals(spaceAvatar.getFileInfo().getName())) {
-      Profile profile = spaceIdentity.getProfile();
-      fileService.deleteFile(spaceAvatar.getFileInfo().getId());
-      profile.removeProperty(Profile.AVATAR);
-      profile.setAvatarUrl(null);
-      profile.setAvatarLastUpdated(null);
-      space.setAvatarAttachment(null);
-      identityManager.updateProfile(profile);
-      space.setAvatarLastUpdated(System.currentTimeMillis());
+    Profile profile = spaceIdentity.getProfile();
+    if (profile != null && profile.getProperty(Profile.FULL_NAME) != null && !profile.getProperty(Profile.FULL_NAME)
+                                                                                     .equals(space.getDisplayName())) {
+      FileItem spaceAvatar = identityManager.getAvatarFile(spaceIdentity);
+      if (spaceAvatar != null && spaceAvatar.getFileInfo().getId() != null
+          && EntityConverterUtils.DEFAULT_AVATAR.equals(spaceAvatar.getFileInfo().getName())) {
+        profile.setProperty(Profile.FULL_NAME, space.getDisplayName());
+        profile.removeProperty(Profile.AVATAR);
+        profile.setAvatarUrl(null);
+        profile.setAvatarLastUpdated(null);
+        space.setAvatarLastUpdated(System.currentTimeMillis());
+        identityManager.updateProfile(profile);
+
+        FileService fileService = CommonsUtils.getService(FileService.class);
+        fileService.deleteFile(spaceAvatar.getFileInfo().getId());
+      }
     }
   }
